TSPostEvaluate: Support for setting a custom post evaluate function (Only added to TSRK for now)
TSPostEvaluate can be used for modifying the solution after the step evaluation. Having TSPostEvaluate enables
the user to solve DAEs with PETSc's explicit methods using a partitioned approach in which the differential
variables are updated by the time-stepping solver while the algebraic variables are updated through user
callbacks in TSPostStage() and TSPostEvaluate(). Another use of TSPostEvaluate is to modify the solution such as
via some filtering. Typically such filtering operations (from Emil) are done in TSPostStep(). However, TSEvent can
modify the solution if it detects an event thus making the action taken in TSPostStep() invalid. So, instead of
doing the filtering in TSPostStep(), the users should rather do it in TSPostEvaluate().