Extract the submatrices
Extract the submatrices from the preconditioning matrix instead of the Jacobian matrix.
The reason for this is that the MatMults for the submatrices (at line 1004, 2035 and 1054 of fieldsplit.c)
involves function evaluations (which is not cheap) when we are using the matrix-free method for inexact Newton.
I have some tests in MOOSE, and this change halves the compute time.