Added DMNetwork example ksp/ex2
This new DMNetwork example is based on ex1 but now the network is
generated with a random algorithm and the user can select the number
of nodes of the network.

Testing larger networks might help us to reveal performance bottlenecks.

Parameters beta and alpha have been selected by trial and error so the
network is connected. But the user can try to select different values.
Different alpha and beta will lead to different clustering, which will
also influence the spectral characteristics of the Laplacian.