When building metis/parmetis make it use a portable random number generator that produces the same result on all sysystems.
It is not 100 percent clear that this option will result in the same partitioning on all
systems but if it does I will be a very happy camper. It is exceedingly annoying that this
is not the default random number generator for metis/parmetis and that it is not a documented feature. We have
wasted endless time with our tests that used metis/parmetis and produced different results
on different systems so we ended up removing the use of metis/parmetis from our nightly tests
which is not a good thing.

Time: 1 hour
Reported-by: endless pain