Enhancements for KSP solve with singular matrices
update KSPSolveTranspose() to utilize the nullspaces attached to the matrix with MatNullSpaceSet()
add test examples that utilizes the singular matrices nullspaces for KSPSolve() and KSPSolveTranspose() with
right and left preconditioning.
Lock vectors in MatNullSpaceCreate() so user doesn't accidently change them when they should not
Add details to KSPSolve() manual page for singular systems.

Thanks-to: Fande Kong <fande.kong@inl.gov>