ViennaCL: Added "row-scaling" preconditioner.
This is a simple Jacobi-like diagonal preconditioner.
The preconditioner computes the norms of each row of the system matrix and uses these values for the diagonal of the preconditioner matrix.
In contrast to PETSc's native Jacobi preconditioner, this preconditioner bypasses a couple of unnecessary host-device transfers (via e.g. PCI-Express) for GPUs.