DMPlex: reduce max iterations for point location.
As I mentioned in a previous commit, there are valid trilinear hexahedra
for which Newton's method will inevitable converge to roots outside the
reference element, even when there is also a root in the reference element.
So, given that we are accepting imperfection from point location (a method
that is not guaranteed to work for all valid element maps), I chose a
maximum number of iterations that seems to be sufficient for even some
hard-case deformed elements.