snes_type test flag for printing entries above threshold
-snes_type test -snes_test_display displays the difference between the
hand-coded jacobian and the finite differences jacobian. This is useful
to debug errors in the jacobian.

However, when the jacobian is too large the errors might be difficult to
spot at a glance. And additional flag:

This pull request implements an additional flag:

-snes_test_display_threshold <threshold>

Where only the entries above the specified threshold will be displayed.

Note: since the current snes_test interface exits with a
PETSC_ERR_ARG_WRONGSTATE, valgrind complains of memory leaks. I'm not
sure how to create a test for this.