Have PETSc install its examples in the share directory
This is incomplete, it puts the entire PETSc directory tree in the share/examples directory tree
but only files in the test/tutorials/output directories. It would be good to have it prune out irrelevent directories
One can run the test suite from this location.

Funded-by: IDEAS
Project: IDEAS/xSDK
Time:  4 hours