Code fixes for TS step restarting (used in ARKIMEX)
i) TSClone() in ARK step-restart was not copying over the event and
trajectory data structures.
ii) When TSDestroy() is called at the end of ARK step-restart, the event
and trajectory structs were getting destroyed.
iii) ARK step-restart calls TSSolve() that resets ts->steps and
ts->total_steps. This caused the filenames created by TSTrajectory to be
overwritten and thus the saved solutions in the file were not matching
with the correct step numbers.