Docs: manual - remove reference to specific MPE event deactivation
It appears that PetscLogEventMPEActivate and PetscLogMPEDeactivate
were removed long ago (they exist in 2.1.0 but not 2.1.6). This edit
assumes that MPE was changed to simply be another logging impl,
and is thus controlled uniformly with other logging (thus,
control individual event logging with the normal interface).

Some other references to those removed functions persist in the other
docs - another commit to follow will remove them.