Test verbosity improved, new implicit rules
Verbosity now handled for tests in the same way as the main library
compilation by using the testquiet command.

Executables now have implicit rules in gmakefile instead of being
explicit in the testfiles file.

Some prototype code is in place for doing implicit rules for running
the scripts, but it is not working.