Enable timed removal of the executable files
Remove executables at the pkg-lang level which is close
to the legacy method.  It does so by having generated rules, so
this could possibly be improved, or perhaps the INTERMEDIATE
will work eventually, but this is progress.

Did some cleanup after Jed's excellent implicit rules which
went off of the counts.

Also, `make cleantest` now regenerates the files.
This seems to be the safest method due to issues mentioned in previous
commit. Basically, the idea of clean is to create a state similar to
after configure, and that includes generating the test targets