gtest: move more generated targets, revamp search
There is no separate 'testsearch' target, just use 'test'.  If a search
argument is given, keep the executables.  If no search argument is
given, run tests by packages and delete executables as we go.  Use

  make -f gmakefile test search=%

to run all tests and keep the executables.  Regardless of which tests
are run, conclude by printing the summary (within the selected set).