VecTagger: absolute implementation and simple backend
The absolute, relative, and cdf backends have a lot of common code for
tagging a single interval of values, so that is constolled by a common
VecTagger_Simple data type.  All the absolute implementation needs to do
is return the stored interval.

WIP: VecTagger absolute implementation

WIP: share a lot of VecTagger code in simple type