Multiple fixes enabling better for loop capability
Enabling args override in the shell script broke the
for loop functionality.  By separating variables that
are in for loops and those that are not, arg override
are allowed even when for loops are present.

Because some for loops cause different outputs and
some produce the same output, we automate the two
extemes.

Also comments are now pulled out of the test specification
and placed at the top of the script.