Enable multiple test from single test block
To split the tests specified into the testblock,
the implementation is done at the testparse.py level
to enable easier debugging and implementation.
For loops and subtests are still specified at the
gmakegentest.py level.  Improvements to parsing
for some error checking were implemented as well.

To test, the changes two examples have been converted:
  src/dm/impls/plex/examples/tests/ex10.c
  src/ksp/ksp/examples/tutorials/ex10.c