Allow users to change the PCSide and NormType after they have been set via the options database
Since the allowed pc sides and normtypes are related the KSPSetFromOptions() would error since it
changed the NormType and then tried to determine the PCSide, but the previously set PCSide was not
compatible with the newly set NormType. Changed the code to allow one to turn off the error checking
during this checking of current (or default) value.

Reported-by: Pierre Jolivet <pierre.jolivet@enseeiht.fr>