Added support to SNESLINESEARCHBT to handle Inf or Nan for larger values of lambda
Force SNESLINESEARCHL2 to set maxstep to the newly computed lambda if the larger lambda produced
Inf or Nan. This is to prevent cycling where the secant step sets lambda large enough to produce
Inf or Nan, the lambda gets cut because of this but the secan step again forces the large lambda.

Reported-by: Till Martens <till.martens@dlr.de>