config: libraries.toStringNoDupes puts LDFLAGS before LDLIBS
This normalized ordering is cleaner/easier to read.  It also fixes what
appears to have been a bug when handling a ["-framework", "libname"]
sequence during library removal.  That is, if we have

  -framework foo -framework foo -lbar

then the former code would have produced

  -framework foo -framework -lbar

while the new code will merely return the first version.  (It's slightly
complicated to remove this duplicate and I doubt it is important.)