Fix flaw with TSSetRHSJacobian() and no TSSetIJacobian() where DM provides the Jacobian matrix
For example src/ts/examples/tutorials/advection-diffusion-reaction/ex5.c
TS would create the Jacobian from the DM in TSGetRHSMats_Private() and then make
a copy of it resulting in unneeded matrix creation and unneeded MatAXPY computations at
each Jacobian computation. MatAXPY() would take up to 80% of time.

The management of Jacobians for the various cases of implicit, explicit, IMEX, with and without RHSJacobian
or IJacobian is excessively convoluted and should be refactored to make the code much simplier and
clearer.

Commit-type: bug-fix
Funded-by: IMEX
Project: IMEX
Time: .7 hours
Reported-by: Hong Zhang <hongzhang@anl.gov>