PetscMath: check for math functions by function pointer resolvability
I had previously tried to solve the issue of hidden symbols in math
libraries being misused in c89.  Trying to set a pointer to the
functio will fail if the symbol is hidden, so this is a better
solution.  I have applied it to other c99 math functions (erf, tgamma)
as well.

Suggested-by: Jed Brown <jed@jedbrown.org>