Make build requirements explicit
Previous logic had builds turned off if there were no tests to be run
for a given configuration.  Now, one has to be explicit on requirements
to build the file.  Two examples of files that had to be updated to
specify the new requirements are sys-tut-ex10f90 and sys-tut-ex9 which
require yaml.