Modified DMNetworkDistribute function call
Previous function call to DMNetworkDistribute would take the DM
object as input and return a pointer to a new distributed DM.

Now we pass the DM by reference, check for the number of processes
and replace the DM object by a distributed version if necessary.

This allows us to avoid the conditional call to the function.