Change TSSetCostIntegrand interface
Originally PETSc creates a sequential vec to hold the integral values for users. This is ok for common senarios where there is just one integral to calculate. When there are multiple integrals (e.g. coming from multiple cost functions), one might want to do it in parallel. The new interface allows users to provide their own vec to support parallelization. If the users do not provide one, a sequential vec will be created.