Allow ell to work for matrices of any size
Previously an error will be thrown out for matrices whose sizes are not multiple of 8. This commit modified the data structure of ELL to support for matrices of any size. Using the exisiting data structure would lead to nonvectorizable code to handle the remainder rows in the last slice. To improve SIMD efficiency, we pad rows of zeros in the last slice when the size is not multiple of 8. Again, the bit array can prevent access to the padded elements for safety.
Funded-by:
Project: PETSc for Theta
Time: 8 hours
Reported-by:
Thanks-to: