New script for analyzing the output of nightlies
A big advantage of the Test Anywhere Protocol (TAP) is it's
simple format -- it is easy to parse.

```exampleslog.py``` analyzes examples that are in the new harness,
and then outputs two new html files:
  * example_summary-sortByPerson.html
  * example_summary-sortByPkg.html
(with the default names).

Usage is:
```
    exampleslog.py -l $PWD -p $PETSC_DIR -o next_example_summary
```

Use:
```
    exampleslog.py -h
```
to see full options.  This has been tested locally on my laptop with the
nightly log files from wget (looked at next from 6/15/2017)

The parser is simple, but the html output took awhile to make it look decent.
Some html boilerplate was moved to htmltemplate.py for ease in programming.
This could be combined with some of the runhtml.py boilerplate to make
development easier.