Enable skip of diff tests when primary cmd fails
If the primary command fails, then the running the
diff test adds a lot of unneeded verbosity.  This
fixes that by checking the error code and SKIPping
the diff test if the command fails.

As part of this, the capturing of error output and
status was changed to be more accurate:
see config/petsc_harness.sh for details.
Currently sys-tests-ex{1,2,3} are failing because
they are meant to capture the error codes.