Enable the testing of errors.
To test the error stream, the test has to be set up
differently to capture the error stream and error
codes correctly.  This is a type of filtering process,
so a new "Error:" code is added to filter.

For example, sys-tests-ex1 uses:
     filter: Error: egrep "(PETSC ERROR)" | egrep "(main|CreateError|Error Created)" | cut -f1,2,3,4,5,6 -d" "
to denote that the error stream is to be tested.