New features: time limit and logging
Enabling a time limit is now enabled.  The default is set in
```config/petsc_harness.sh``` (at 60 seconds) but may be
overwritten at the shell level with the ```-t intfoo``` flag,
and at the makefile with the ```TIMEOUT=intfoo``` option.
Individual tests can be extended with the
```timeoutfactor: intfoo``` in the test block.

Although the pipestatus variable in bash can get the
equivalent of ```$?``` for piped processes, such a thing does
not exist for ```$!```.  To overcome this, a new script
```config/watchtime.sh``` is used for enforcing the timeout.
Because the timeout requires separating the filter from
execution of the job, pipestatus is no longer used and the
filter step is done separately.  The implementation in
```petsc_harness.sh``` is explained by these.  Also,
interrupting the tests  with ```CTRL-C``` means you
will see a kill status appear at the terminal as this dies.

Logging is also enabled.  Previous attempts at doing this
purely at the ```gmakefile``` level were unsatisfactory
because of pipe buffering issues.  Moving the
```| tee -a ${testlogfile}``` command to
```petsc_harness.sh```
level gives acceptable performance.  The logfile is:
```${PETSC_ARCH}/tests/examples-${PETSC_ARCH}.log``` and
```grep -v ^ok ${PETSC_ARCH}/tests/examples-${PETSC_ARCH}.log```
is very useful.