Allow KSPPREONLY and KSPFETIDP to "support" any KSPNormType
this prevents errors when the user has hardwired a particular KSPNormType
that is not relevant for these methods.

Also improved documentation regarding the fact that KSPPREONLY does use
KSP convergence options and KSPFETIDP automatically copies options supplied to the
outer KSP to the inner one.

A few fixes for use of VecGetArrayRead() because Barry's work flow is terrible.
These are unrelated to the previous updates.

Commit-type: style-fix, feature, documentation
Time: .4 hours
Reported-by: Jed Brown <jed@jedbrown.org>