DMNetwork: Added DMNetworkGetComponent to get the component data directly.
The component data used to be accessed by obtaining the component data array first
through DMNetworkGetComponentDataArray and then by getting the offset for a given point
via DMNetworkGetComponentKeyOffset. Added a new function DMNetworkGetComponent through which
the component data can be accessed directly without having to expose the component data array
or the offset.