Remove redundant info from gmakefile.
Per discussion, gmakefile.examples has been changed to
gmakefile.test.  It is now included in gmakefile to enable
the removal of redundant information.  Tests on whether
gmakefile or gmakefile.test is invoked is used to print
out the appropriate help information for that file (i.e.,
do not print out test info for basic compiles).