Add feature that Jed requested in PR #739
It is difficult to rerun failed tests partially because of the mapping
of makefile targets (which are oriented to the scripts that are being
run), and the test names that are actually reported (which are
oriented around individual tests being run, including for loop names,
test sets, etc.).  A set of regex's are used to provide the inverse
name mapping that is implicit in testparse.py to provide a command
to enable people to copy and paste the line.