MATELL: Added AVX implementation for MatMult()
Updated the MATELL datastructure as follows:
Unused entries are set to the last valid column (column array) and zero (value array) during MatAssemblyEnd().
This allows for the use of AVX intrinsics instead of AVX2 for MatMult().
Also, my experiments indicate that any masking is detrimental to performance on Xeon CPUs in this context, so AVX-routnies are preferable over AVX2-routines anyway.