PetscMallocA: implement the Malloc[1-7] suite using variadic function
The macros had become increasingly complicated and were opaque to debug.
It also prevented run-time choice of coalesced malloc, which is now
available via -malloc_coalesce which is useful to debug memory leaks due
to mismatched PetscMalloc3/PetscFree2 and the like.  This implementation
dispatches to a single variadic function that handles allocation for
both PetscMalloc[1-7] and PetscCalloc[1-7], as well as a matching
case for freeing memory.