PFLOW: Bug-fix in parser.
Some MATPOWER data files also have the field mpc.bus_name that was causing
the parser to break. This was because of strstr(line,"mpc.bus") code that
searched for the mpc.bus tag in the line. With "mpc.bus_name" field present,
the parser was incorrectly reading that as well.