PetscMallocA: update macro expansions to cast numbers to (size_t)
Rationale:

  1. Silence -Wconversion warnings.
  2. Your code may be broken because of integer promotion rules.

As an example of (2), imagine that you are running on a 32bit arch with
size_t 32 bits and long long 64 bits. Then

  unsigned long long m1 = 10;
  unsigned long long m2 = 10;
  double *p,*q;
  PetscMalloc1(m1,&p,m2,&q)

Then the macro will compute the product m2*sizeof(double), the result is
then unsigned long long, but PetscMallocA is expecting size_t (because
of va_arg(Argp,size_t)).

Thanks-to: Lisandro Dalcin <dalcinl@gmail.com>