PetscViewerDestroy_ASCII: clean up stdout and stderr attributes from MPI_Comm
Since -log_view is processed after PetscObjectRegisterDestroyAll() in
PetscFinalize(), it can attempt to access attributes from
PetscViewerASCIIGetStdout().  When there are no references to the inner
communicator, it will be freed no later than
PetscObjectRegisterDestroyAll() and thus all attributes become unset.
But if there is any reference to the inner communicator, as in the
example below with the dummy PetscContainer (the type does not matter),
it remains after PetscObjectRegisterDestroyAll() and thus any stale
attributes remain accessible.

  #include <petscviewer.h>

  int main(int argc,char **args)
  {
    PetscErrorCode ierr;
    PetscViewer    view;
    PetscContainer container;

    ierr = PetscInitialize(&argc,&args,(char*)0,NULL);CHKERRQ(ierr);
    ierr = PetscContainerCreate(PETSC_COMM_WORLD,&container);CHKERRQ(ierr);
    ierr = PetscViewerASCIIGetStdout(PETSC_COMM_WORLD,&view);CHKERRQ(ierr);
    ierr = PetscFinalize();
    return ierr;
  }

Reported-by: Stefano Zampini <stefano.zampini@gmail.com>