Fix MatConvert error when used with hypre
SNES ex56 gives an error when ran with -pc_type hypre -mat_block_size 3. It is because MatSetLocalToGlobalMapping checks the block size and errors out when *M and mat have different values of bs. However it is unnecessary to impose same bs for MatConvert.

Funded-by:
Project: PETSc for Theta
Time: 0.5 hour
Reported-by:
Thanks-to: