Dependency fixes.
The actual extra dependencies from the `depend:`  keyword
were not being written out explicitly to testfiles.  This
now does so.

The new rules for compiling test executables then requires that all of
the dependencies be added to the link line which changed the rule from
using `$< $(PETSC_LIB)` to a `$^`.