PCTelescope: Rename DMView_DMDAShort functions
This is to maintain the convention that for interface functions,
appending an underscore and a single word after the name of the
the function represents an implementation of that function. Thus,
before this change, DMView_DMDAShort would imply an implementation of DM called
DMDMDAShort (which doesn't exist) which provides custom DMView
logic.