Adding alt file replace functionality
Adding ALT=1 to the gmakefile.test arguments will enable passing the -M
to the shell harness which then passes it to petsc_diff which will then
replace the alt file.  Should use with caution if multiple alt files are
present.

A bug was found while enabling this for alt files.  I used a || operator
instead of a && operator.