Reset the random seed for each partition that PTScotch may perform so as to get consistent results.
Note this does not resolve the reason for the difference between PetscPartitioner and MatPartitioning

I found the problem! In plexpartition.c (line 1750) there is this horrible special case where if it determines a single process owns the entire graph (in this case MPI_Comm_rank 0) it calls a different sequential partitioner instead of the parallel preconditioner. MatPartitioning does not handle such special cases differently than others so, of course, uses the parallel partitioner and generates a different partition. @Matthew Knepley We could introduce this special case into the MatPartitioning for PTScotch and thus reproduce the PetscPartitioner results or we could say that handling the special case is goofy and live with the difference. Or we could remove the special case from PetscPartitioner and get identical results for both.

Commit-type: portability-fix, testing-fix