Man page processing: add class implementations
This looks for structs of the form _p_XXX and adds links
to the man page for XXX.html. For example, you can go the DM.html man
page and jump from there to dmimpl.h to look at _p_DM.

Note that this doesn't cover things like DMDA, which has additional
data in a DM_DA struct. A future improvment could add those sorts
of links as well, but a usable workaround is to note that dmdaimpl.h is
directly included in the file where DMDA is defined, so one can at
least get there by clicking more.