Improvements for the install
The test harness has to operate in 3 modes:
  1. Normal development mode
  2. Test install from source dir (make allgtests)
  3. Test install from install dir
See config/gmakegentest.py documentation for how these differs.
This sorts out the paths for all 3 cases.

A big change is making gmakefile.test use absolute paths
rather than the relative paths used originally and in the normal
gmakefile.  This simplifies a lot of the development, including the
previous method of doing a lot of modifications for the installs.

`make test` is still not hooked up to the new test system so testing was
done with allgtest.