DMPlexSNES: caching of new geometry constructs
The old paradigm cached one geometry object for all cells or all faces,
an approach that doesn't take into account that different quadratures
may require different geometry objects.  Now the geometry object
is composed with the index set over which the geometry ranges, and the composition
string include the object id of the quadrature.  To take advantage of
this caching, we have to be careful not to reconstruct index sets on the
fly when they are known a priori.  In particular, the hybrid bounds
have been encapsulated as a label, which makes them accessible as ISs.