make: allow -j options in top-level makefile
Removes explicit mention of MAKE_NP and MAKE_TEST_NP which are now just
defaults and are superseded by explicit MAKEFLAGS which in turn are
superseded by command line -j.

Remove chk_makej which does not work with GNU Make due to "special"
handling of MAKEFLAGS in the context of these tests.