Removing pure test harness test
This test occasionally takes a long time although it seems to not be
reproducable.  Since this test is no longer needed -- we have enough
other *real* examples that we don't have to test the test harness
itself, just removing it seems more useful than trying to debug the odd
behavior.