Allow to override build date
in order to make builds more reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

Also dont store the buildhost name if the user desires
a reproducible build.

Without this patch, petscmachineinfo.h and the resulting .so file
would differ for every clean build