PetscFEGeomComplete(): correct normal vector in 2D.
We get the normal vector from the unit basis vector added to complete
the Jacobian in a way that makes the determinant positive.  In 3D, this
is fine, because by convention the facets are oriented with outward
normal, but in 2D they are not, so we have to invert the sign.