PETSCSPACETENSOR: A generic tensor product of spaces.
The spaces do not have to be the same.
The spaces do not have to be one dimensional.
Eventually, e.g., we could use this on prisms.
For the most common case, we should probably
Convert to a simple PETSCSPACEPOLYNOMIAL at PetscSpaceSetUp()

wip: PETSCSPACETENSOR

wip: PETSCSPACETENSOR

wip: PETSCSPACETENSOR

wip: PETSCSPACETENSOR

WIP: PETSCSPACETENSOR

wip: PETSCSPACEPOLYNOMIAL

wip: PETSCSPACETENSOR