Optimize local scatter when it is made of multiple copies
Old code optimized local scatter when it was actually a copy. We extend it to handle cases
when local scatter is made of multiple copies.

One particular application of this optimzation is when we call DMGlobalToLocalBegin/End to
copy a global vector to a local vector created by DMCreateGlobalVector or DMCreateLocalVector
(see src/ts/examples/tutorials/advection-diffusion-reaction/ex5.c). The global vector is
contiguous in memory, but interior part of the local vector is only piece-wise contiguous in
memory. Therefore, multiple copies are needed.