Do VecScatterLocalOptimizeCopy after VecScatterRemap
VecScatterRemap() is used to translate a DMGlobalToLocal vecscatter context to a
DMLocalToLocal vecscatter context. The remapping may invalidate old optimizaitons or
enable new optimizations. So we call VecScatterLocalOptimizeCopy after VecScatterRemap
to start optimizations from scratch

VecScatterLocalOptimizeCopy() was VecScatterLocalOptimizeCopy_Private(). This commit
made the function not private anymore so we changed its name.