Update timer reporting and minor reporting fix
Realized that loops and multiple tests were not having their time
recorded properly when using the time command.  This now appends
to the file and then uses awk to total all of the times.  Went
ahead and switched to user+sys time as well since that seems
more accurate.

Also, fix for the telling the user which tests to rerun when failing
when there is a label_suffix (e.g., snes_tutorials-ex12_restart-b).
Reported by @satish