PCFieldSplit: Interface to set -pc_fieldsplit_detect_saddle_point programmatically
Adds a new boolean data field for this setting, and a setter and getter.

Removes a call to get the option outside of PCSetFromOptions_FieldSplit.

In PCSetFromOptions_Fieldsplit(), checks the flag and then immediately
calls PCFieldSplitSetDetectSaddlePoint(), which then set the split type
and SchurPre type, as before.

Wraps setting PC_FIELDSPLIT_SCHUR_PRE with the existing API.

Updates documentation, in particular removing a statement that negative diagonal
entries trigger a split (when only zero diagonal entries do).