Renamed ex10 to ex72 to make room for the ex10 used in PETSc manual
src/ksp/ksp/examples/tutorial/ex10.c is mentioned in the profiling chapter of the manual but
the code is out of sync with the manual. So we renamed the existing ex10 to make room
for a new ex10. In this way, people using old PETSc manuals can still be good.