DMDA VTK Viewer: write multi-dimensional data to .vtr and .vts
Change the behavior to output a multi-dimensional data field for
multi-dof DMDAs.

The requirement that all vectors written to the file come from the same
DM is relaxed: compatible DMDAs (the same grid, different dof) are now
allowed.

This makes it simpler to do things like dump a pressure field and
velocity field (must be 3 dof even in 2d, with paraview) and be able to
quickly load it.