VTK Viewer: add flag to skip DM check PetscViewerVTKAddField(), compatibility check for DMDA, PetscViewerVTKGetDM()
In the DMDA case, this allows for compatible, yet different, DMDAs to be attached to the provided Vecs.

This is done as the fields are added, by using a new
PetscViewerVTKGetDM() function to check for compatibility,
with the help of DMGetCompatibility(). The new flag for
PetscViewerVTKAddField() is used to skip the existing, more stringent
requirement that all fields must come from vectors generated from the
same DM.