SNES: Backporting 3.9.x API, reintroducing threshold print.
Barry's fixes to SNESTestJacobian (accidentally?) removed the possibility
of only printing a threshold-version of the diff'd Jacobian matrices.
This commit adds this option back in, using the old -snes_test_jacobian_display_threshold.
For backwards compatibility, also -snes_test_jacobian_display is reintroduced.

Eventually, we want to unify the runtime options used here.
This needs to happen in a follow-up commit that does not get merged to the 3.9.x-version of 'maint'.