Tutorial examples: add test cases to use with hands-on tutorials
Add test cases for use with hands-on tutorials.

These are all named "exampleName_tut_###".

They are atypical entries for tests, in that the purpose is to keep
up-to-date reference output (*tut_###.out files) for the tutorials,
not to check for regressions in library functionality.

Note that many tests are ignored in single precision (since the
tutorials will presumably be performed in double precision), and more
verbose output is included that for a typical test.

Note that with commands included -log_view, aggressive filtering is
perfoperformed (skipping almost all of the output).