PetscFileRetrieve(): Suppress output to STDERR when downloading a file.
curl by default prints a progress report to STDERR when downloading a file.
When used from PETSc, we don't want to pollute STDERR with such an internal detail.
Thus, passing flags to curl to only output to STDERR if there is an actual error.

See also https://superuser.com/questions/173209/curl-how-to-suppress-strange-output-when-redirecting