Implemented an AMG that uses Hypre to coarsen matrix
and generate a sequence of coarse matrices. These coarse
matrices are used to setup a PETSc PCMG.

There are two benefits:

(1) We can use any PETSc preconditioners such as ASM, BJacobi as
level smoothers (solvers). For some applicaitons, the combination
of ASM and AMG works great while only AMG or AMS does not work at all.

(2) For multicomponent problems, we could just coarsen one submatrix
associated with one particular component. In this way, the setup of the
preconditioner is significantly improved. One typical use case is
neutron transport equations. There are many variables on each mesh
vertex due to the discretization of angle and energy. Each variable,
in fact, corresponds to the same PDEs but with different material
properties.