config executeShellCommand: support chained commands and non-shell
The present approach of concatenating strings is problematic if paths
contain spaces or special characters.  We should always use lists
instead of strings, but I don't have the patience to convert all call
sites right now.  I have converted a representative sample, however,
to demonstrate that the approach works.  This involved a new function
Script.executeShellCommandSeq() that takes a sequence/list of commands
to be run.

A legacy pattern in BuildSystem has been to use chained sequences of
the form

  cd to/dir && run command

instead of using the cwd keyword argument.  This commit fixes many
such instances to use the cwd argument.