ierr for PetscErrorf is intent(in)
intent(in/out/inout) specifies the read/write/read+write access for
arguments of functions and subroutines.
Therefore, a constant expression is not allowed for intent(out).
ierr is usually intent(out) as it contains the error code on return of a
function with PetscErrorf being the only exception as it reads ierr and
returns the matching error message.