MRK: add multirate suport for more than 2 partitions
The key idea is to allow for slow-fast paritioning recursively. For example, the first level could consist of a slow part and a fast part. Then the fast part can be futher split into a slow part and a fast part. Each subts has its own rk context, but share as many working vectors as possible.