gmakefile: distinguish MODDIR and TESTMODDIR
We don't actually use the same rules for building Fortran library
sources and Fortran tests, but when a target-specific variable is set
for a test, that target-specific definition is in scope for all
prerequisites, including the library.  As a result, if library modules
were built as a result of executing

  make -f gmakefile test

then MODDIR would point at the directory containing the target object
file instead of the necessary include directory.  This commit leaves
MODDIR intact by defining TESTMODDIR to be used for the tests.