Fix error with PetscComplex when PETSc is compiled with C and used in C++
If PETSc is compiled with a C compiler then PetscComplex is typedefed to
a C type. When petscmath.h is used in a C++ file, PetscComplex is
initialized as if it were a std::complex independent of what it actually
is. This is fixed by checking if PETSC_HAVE_CXX_COMPLEX is defined
before initializing PetscComplex.