Sketch an "owned" patch constructor.
This has several uses. First of all, in serial it allows us to verify
the implementation of SNESPATCH (by comparing it to Newton/LU or
whatever). Second, this begins to address the functionality offered by
PCASM/SNESNASM, but in a more general way.

At the moment, in the patch code we compute the sparsity pattern of the
matrices using a PetscBT. This doesn't work for moderately sized
patches, as the comment in PCPatchCreateMatrix_Private indicates. This
should be fixed.

The convergence of this might be greatly improved by keeping some
degrees of freedom in the overlap; we should add options to allow the
user to control this, also. (This might entail changing the name.)