Rename PC_PATCH_OWNED to PC_PATCH_PARDECOMP.
Still not entirely settled on the name. However, "owned" is misleading
as it will soon incorporate overlap, so it's not just "owned". Since
it's constructing the domain decomposition from the solver from the
parallel decomposition, I want to get "parallel" in there somewhere.
-pc_patch_construct_type paralleldecomposition seems a mouthful,
so I thought let's go for -pc_patch_construct_type pardecomp for now.