PCBDDC: add lumped variant, aka M_1
This variant, also known as M_1 in the literature, skips extension of
balanced interface values to the interior by solving a Dirichlet
problem.  It gives up a factor of H/h in condition number relative to
Dirichlet (standard) BDDC, but may be appropriate for small
subdomains, inexact subdomain solvers, and/or as an ingredient in
composite preconditioners.

Based-on-patch-by: Stefano Zampini <stefano.zampini@gmail.com>