Docs: manexamples: more-specific matching for counting example links
The manexamples target was counting the number of example links alredy
added to a man page by counting the occurences of "tutorials", which
lead to an incorrect count on pages (e.g. KSPSolve) which happened to
mention this word. This restricts to counting lines which start with a
link opening tag and contain /tutorials/.

Reported-by: Dave A. May <dave.mayhem23@gmail.com>