PetscCheckPointer: speed up check by using existing signal handler
Registering the special signal handler dominates the cost of these
checks.  Here we use the existing (default) signal handler.  This has
a race condition in the presence of threads, though SEGV is typically
a "bad enough" condition that threads shouldn't race for it.  In any
case, all the signal() based error handling is also ill-defined for
threads.  To make this correct and non-racy with threads, we'll need:

1. use thread-local storage for the flag PetscSegvJumpBuf_set

2. use sigaction instead of signal (requires overhauling signal.c)

This commit changes behavior in case a user has registered their own
signal handler, but that change should only be observable when a SEGV
is raised, and may be avoided with -check_pointer_intensity 0.