Add a ScatterMode arg in cuda vecscat to select to/from context
The old code VecScatterInitializeForGPU() initializes the pointer (PetscCUDAIndices*)&inctx->spptr) based
on an input ScatterMode before VecScatterBegin() is called.

If a vecscatter context is firstly used for a SCATTER_FORWARD, and secondly used
for a SCATTER_REVERSE, there will be an error. Since in the second VecScatter,
it uses out-of-date (PetscCUDAIndices*)&inctx->spptr)

The solution is "do not prematurely consider ScatterMode when building (PetscCUDAIndices*)&inctx->spptr). Instead,
select correct to/from until VecScatterBegin() is called"